Split Graph into multiple files, implement selection
