+++ web/.eslintrc
@@ -19,6 +19,7 @@
     "vue/multi-word-component-names": "off",
     "vue/valid-v-slot": "off",
     "@typescript-eslint/no-explicit-any": "off",
-    "@typescript-eslint/explicit-function-return-type": "off"
+    "@typescript-eslint/explicit-function-return-type": "off",
+    "@typescript-eslint/no-non-null-assertion": "off"
   }
 }
