+++ web/src/views/analysis/cluster.vue
@@ -94,6 +94,7 @@
                 :clustering="clustering"
                 :height="350"
                 :node-size="8"
+                :show-singletons="false"
                 node-tooltip
                 node-clickable
                 @click:node="onNodeClick"
